refactor: 세션 상수 관리 클래스 이름을 변경
`AuthConst` -> `SessionConst`
