+++ src/main/java/com/programmers/board/constant/SessionConst.java
@@ -4,6 +4,6 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public abstract class AuthConst {
+public abstract class SessionConst {
     public static final String LOGIN_USER_ID = "userId";
 }
