+++ src/test/java/com/programmers/board/controller/LoginControllerTest.java
@@ -20,7 +20,7 @@ import org.springframework.mock.web.MockHttpSession;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 
-import static com.programmers.board.constant.AuthConst.LOGIN_USER_ID;
+import static com.programmers.board.constant.SessionConst.LOGIN_USER_ID;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.InstanceOfAssertFactories.STRING;
 import static org.mockito.ArgumentMatchers.any;
